+++ eras/dijkstra/impl/testlib/Test/Cardano/Ledger/Dijkstra/ImpTest.hs
@@ -80,6 +80,7 @@ instance ShelleyEraImp DijkstraEra where
   modifyImpInitProtVer = conwayModifyImpInitProtVer
   genRegTxCert = dijkstraGenRegTxCert
   genUnRegTxCert = dijkstraGenUnRegTxCert
+  delegStakeTxCert = conwayDelegStakeTxCert
 
 instance MaryEraImp DijkstraEra