+++ eras/shelley/impl/CHANGELOG.md
@@ -11,6 +11,8 @@
 * Added `genUnRegTxCert` to `ShelleyEraImp`
 * Added `shelleyGenRegTxCert`
 * Added `genRegTxCert` to `ShelleyEraImp`
+* Added `delegStakeTxCert` to `ShelleyEraImp`
+* Added `expectStakeCredRegistered`, `expectStakeCredNotRegistered`, `expectDelegatedToPool`, `expectNotDelegatedToPool` to `ImpTest`
 
 ## 1.17.0.0